requirements.txt
numpy
scipy
networkx
scikit-fuzzy
gradio
google-generativeai
pillow
rapidfuzz